% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_counts.R
\name{draw_counts}
\alias{draw_counts}
\title{Draw counts from signal or noise based on logical matrix 'mat'}
\usage{
draw_counts(size_sig, size_bg, mu_sig, mu_bg, mat, seed = NULL)
}
\arguments{
\item{size_sig}{size parameter signal}

\item{size_bg}{size parameter background}

\item{mu_sig}{mu signal}

\item{mu_bg}{mu noise}

\item{mat}{logical matrix}

\item{seed}{random seed}
}
\value{
counts matrix
}
\description{
Draw counts from signal or noise based on logical matrix 'mat'
}
