% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_signal.R
\name{logimat}
\alias{logimat}
\title{Create logical matrix for singlets, doublets and negatives}
\usage{
logimat(ngroups, nsinglet, ndoub, nneg)
}
\arguments{
\item{ngroups}{number of biological samples}

\item{nsinglet}{vector for number of singlets in each group}

\item{ndoub}{number of doublets}

\item{nneg}{number of negatives}
}
\value{
logical matrix
}
\description{
Create logical matrix for singlets, doublets and negatives
}
